Tweak grid Tweaks to the grid, based on the fact that "tablet" as a breakpoint for this particular design isn't great.
To fix this, the "desktop" has been redefined at 925px, based on looking at the
content we have in nhs-design.herokuapp.com - this is the start of a more
content-out approach, rather than predefining "mobile", "tablet", "desktop".
While these are useful to have as large general queries, we're going to want
to be more component led than this.

Also added $measure: 40em; for comfortable reading width. This is used as a
reading-width class when marking up grids.