Remove content_block component
The content block component was used as a way to wrap markdown
blocks in divs that give the content some different visual presentation.

We can't be this specific about which blocks contain that presentation
when the content comes from an API so we're handling when to use those
surrounding elements within the template itself.