Remove two column layout in favour of 1 column component stream
We stack our components currently based on the needs of that
symptom or condition. The use of the second column isn't consistent
with that. Therefore we're dropping back to a singular column that
also contains a max width restriction for blocks of text.

The idea of moving certain components out of their current
order based on screen width doesn't feel very consistent.

Moving forward the side column may be re-introduced but for
genuine supporting content.