Ensure all in-page anchor links have tracking attribute
We currently use a custom data attribute to track in-page anchor
links.

This was being added when using the section_nav component but gets
missed when manually adding anchor navigation to the local header.

This temporarily adds the correct attribute to the existing navigation
items but will need a better long term fix. This fix may well be applied
at the publisher/content store end.