Full stop position inside links
Links had previously been including a full stop/period when the link was
at the end of the sentence or the whole sentence. Latest guidance is
to not include a full stop at all if the whole sentence is a link
and to move the full stop outside of the link text if part of a
set of sentences.