Remove `section_list` and `section_nav` components
In order to simplify some of the components used to build pages we're
removing the need for the section_list component and also therefore
the section_nav.

The only example use of this component is currently the rashes page
which is being revisited from a design perspective to address issues
with in-page navigation and multi sectioned, single page content.

Once we understand the use better we may re-introduce a component
that caters for sectioned content within a page and dynamically builds
out the navigation.