Remove the need for `article_list` and `article_nav` components
Similar to PR #311 this consolidates the use of some of our
components on the frontend. We were previously using `article_*`
components to render parts of the surgery guide. This PR allows the
same visual representation to be achieved using markdown and some
added attributes.

The downside is that we are now using more instances of class names
within markdown to render things like in-page anchor navigation. This
means it is harder to update the structure to these without having to
edit content and could be dangerous moving forward.

We need to find a more robust way to handle internal page navigation
and remove the added overhead of adding classes and attributes to
text content for presentation and analytics tracking.