Switch font to Frutiger across the board
The brand font for the NHS is Frutiger so we should use this if
possible. We were previously testing FS Me but have no conclusive
evidence that it is any more legible so cannot justify its use over
the brand font.

This removes all uses of FS Me that was hosted on a separate
domain. The functionality might be useful in the future if there
is a need to serve the site font from a different domain.